Refactored away a lot of code in favor of design patterns.
1. Notes can be created from a NoteFactory, of which the NoteFactory
will generate an appropriate default identifier for it.
2. There is a map for a NoteManager, one for each File (although this
uses the relative file name and not absolute, and will change in the
future). It is implemented as a Singleton, and so, given the actual File
Name, the user may obtain the NoteManager for that given file. This
NoteManager, as it's name implies, keeps track of all notes for that
given file.
3. Added a new NoteDescriptorPanel which will be used to describe the
state of the currently selected note (I.E, display it's properties) and
will be modifiable. This will, of course, act as the Front-End for the
Controller (which has not been implemented yet) which will manipulate
both the Model (Note + NoteManager) and View (GlossableTextArea ->
NoteTextArea).

Relatively satisfied with these changes.