Finished the bare-bones of the template support. Currently, it looks for a file called "template.json" which much be located in the ClassPath (root of the project), and is parsed into a tree data structure.
The displayed templates can be applied by double clicking on them, which
will alter their message and highlight color. This should help make it
very easy with applying these. In the future, I plan on adding some
shortcut keys which the user may be able to define as their favorites,
but I'm not sure its worth it.

I'll also be refactoring soon, as its getting a bit difficult to add all
of these extensions in rapid succession.